# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Remont API
  version: 1.0.0
paths:
...
  /api/v1/users/:
    get:
      tags:
      - users
      summary: Read Users
      operationId: read_users_api_v1_users__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserOut'
                type: array
                title: Response Read Users Api V1 Users  Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - users
      summary: Create User
      operationId: create_user_api_v1_users__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/users/{id}:
    patch:
      tags:
      - users
      summary: Update User
      operationId: update_user_api_v1_users__id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/users/{id}/reset-password:
    post:
      tags:
      - users
      summary: Reset Password
      operationId: reset_password_api_v1_users__id__reset_password_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/zaprafka/start:
    post:
      tags:
      - zaprafka
      summary: Start Zaprafka
      operationId: start_zaprafka_api_v1_zaprafka_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZaprafkaCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZaprafkaOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/zaprafka/{zaprafka_id}/finish:
    post:
      tags:
      - zaprafka
      summary: Finish Zaprafka
      operationId: finish_zaprafka_api_v1_zaprafka__zaprafka_id__finish_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: zaprafka_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Zaprafka Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZaprafkaOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UserCreate:
      properties:
        username:
          type: string
          title: Username
        full_name:
          type: string
          title: Full Name
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
          default: true
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - full_name
      - role
      - password
      title: UserCreate
    UserOut:
      properties:
        username:
          type: string
          title: Username
        full_name:
          type: string
          title: Full Name
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
          default: true
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
      - username
      - full_name
      - role
      - id
      title: UserOut
    UserUpdate:
      properties:
        active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ZaprafkaCreate:
      properties:
        machine_id:
          type: string
          format: uuid
          title: Machine Id
        start_date:
          type: string
          format: date-time
          title: Start Date
        template_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Template Id
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      type: object
      required:
      - machine_id
      - start_date
      title: ZaprafkaCreate
    ZaprafkaOut:
      properties:
        machine_id:
          type: string
          format: uuid
          title: Machine Id
        start_date:
          type: string
          format: date-time
          title: Start Date
        template_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Template Id
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        id:
          type: string
          format: uuid
          title: Id
        end_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Date
        status:
          type: string
          title: Status
      type: object
      required:
      - machine_id
      - start_date
      - id
      - status
      title: ZaprafkaOut
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: api/v1/auth/login